# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs Gudangin",
        "version": "1.0.0"
    },
    "paths": {
        "/api/barang": {
            "get": {
                "tags": [
                    "Barang"
                ],
                "summary": "Get Barang ",
                "description": "Get Barang ",
                "operationId": "GetBarang",
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/barang/{id}": {
            "get": {
                "tags": [
                    "Barang"
                ],
                "summary": "Get Barang Details",
                "description": "Get Barang Details",
                "operationId": "GetBarangDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Barang",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}